Fixed some GCC warnings

+++ SDLut/src/SDLEvent.cc
@@ -32,7 +32,7 @@ namespace RAGE{
 	std::map<Event::Type, SDL_EventType>::iterator itEnd = EventTyperage2sdlmap.end();
 	for (; it != itEnd; ++it)
 	{
-		assert((*it).first >= 0 && (*it).first < result.size());
+		assert((*it).first >= 0 && (*it).first < static_cast<int>(result.size()));
 		result[(*it).first] = (*it).second;
 	}
 }
