+++ SDLut/test/SDLtestGL.cc
@@ -124,7 +124,7 @@ public:
 
 	  // Logo Faces
 		glBegin (GL_TRIANGLES);
-		  for(int i=0;i<sizeof(face_indices)/sizeof(face_indices[0]);i++)
+		for(int i=0;i< (int) (sizeof(face_indices)/sizeof(face_indices[0]));i++)
 		   {
 		   for(int j=0;j<3;j++)
 			{
